@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:messenger/models/country_code.dart';
 
 class CountryDropDown extends StatelessWidget {
-  final List<CountryCode> listOfCodes;
+  final List<CountryCode>? listOfCodes;
 
-  final CountryCode value;
-  final Function(CountryCode) onChanged;
-  const CountryDropDown({Key key, this.listOfCodes, this.value, this.onChanged})
+  final CountryCode? value;
+  final Function(CountryCode?)? onChanged;
+  const CountryDropDown({Key? key, this.listOfCodes, this.value, this.onChanged})
       : super(key: key);
 
   @override
@@ -14,11 +14,11 @@ class CountryDropDown extends StatelessWidget {
     return Center(
       child: Container(
         child: DropdownButton<CountryCode>(
-          value: value ?? listOfCodes[0],
+          value: value ?? listOfCodes![0],
           underline: SizedBox(),
           elevation: 0,
           isExpanded: false,
-          items: listOfCodes
+          items: listOfCodes!
               .map(
                 (e) => DropdownMenuItem<CountryCode>(
                   value: e,
@@ -30,10 +30,10 @@ class CountryDropDown extends StatelessWidget {
                         Container(
                           height: 18,
                           width: 18,
-                          child: Image.asset(e.flagUri),
+                          child: Image.asset(e.flagUri!),
                         ),
                         SizedBox(width: 5),
-                        Container(child: Text(e.dialCode))
+                        Container(child: Text(e.dialCode!))
                       ],
                     ),
                   ),
