@@ -5,7 +5,7 @@ import 'package:messenger/models/user.dart' as model;
 part 'contacts_model.g.dart';
 
 abstract class PhoneContacts {
-  final Contact contact;
+  final Contact? contact;
 
   PhoneContacts(this.contact);
 }
@@ -14,8 +14,8 @@ class RegisteredPhoneContacts extends PhoneContacts {
   final Contact contact;
   final model.User user;
   RegisteredPhoneContacts({
-    @required this.contact,
-    @required this.user,
+    required this.contact,
+    required this.user,
   }) : super(contact);
 
   Map<String, Map<String, dynamic>> toMap() {
@@ -32,22 +32,22 @@ class RegisteredPhoneContacts extends PhoneContacts {
 
     return RegisteredPhoneContacts(
       contact: Contact.fromMap(
-        newMap['contact'],
+        newMap['contact']!,
       ),
       user: model.User.fromMap(
-        newMap['user'],
+        newMap['user']!,
       ),
     );
   }
 }
 
 class UnRegisteredPhoneContacts extends PhoneContacts {
-  final Contact contact;
+  final Contact? contact;
   UnRegisteredPhoneContacts({this.contact}) : super(contact);
 
   Map<String, Map<String, dynamic>> toMap() {
     return {
-      'contact': contact.toMap().cast<String, dynamic>(),
+      'contact': contact!.toMap().cast<String, dynamic>(),
     };
   }
 
@@ -56,7 +56,7 @@ class UnRegisteredPhoneContacts extends PhoneContacts {
         .map((key, value) => MapEntry(key, Map<String, dynamic>.from(value)));
     return UnRegisteredPhoneContacts(
       contact: Contact.fromMap(
-        newMap['contact'],
+        newMap['contact']!,
       ),
     );
   }
@@ -67,6 +67,6 @@ class HivePhoneContactsList extends PhoneContacts with HiveObject {
   @HiveField(0)
   List<List<Map<String, dynamic>>> phoneContacts;
   HivePhoneContactsList({
-    @required this.phoneContacts,
+    required this.phoneContacts,
   }) : super(Contact.fromMap(phoneContacts[0][0]));
 }
