@@ -2,22 +2,22 @@ import 'package:flutter/foundation.dart';
 import 'package:intl/intl.dart';
 
 class Message {
-  final String message;
+  final String? message;
   final DateTime timeOfMessage;
-  final String senderID;
-  final String receiverID;
-  final String chatID;
-  final String messageType;
-  final String messageID;
+  final String? senderID;
+  final String? receiverID;
+  final String? chatID;
+  final String? messageType;
+  final String? messageID;
 
   Message({
-    @required this.message,
-    @required this.timeOfMessage,
-    @required this.senderID,
-    @required this.chatID,
+    required this.message,
+    required this.timeOfMessage,
+    required this.senderID,
+    required this.chatID,
     this.receiverID,
-    @required this.messageType,
-    @required this.messageID,
+    required this.messageType,
+    required this.messageID,
   });
 
   Map<String, dynamic> toMap() {
