@@ -3,8 +3,8 @@ import 'dart:io';
 import 'package:image_picker/image_picker.dart';
 
 class MessengerImagePicker {
-  static Future<File> pickeImage() async {
-    File _imageFile;
+  static Future<File?> pickeImage() async {
+    File? _imageFile;
     try {
       var pickedFile =
           await ImagePicker().getImage(source: ImageSource.gallery);
