@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:messenger/customs/error/error.dart';
 import 'package:messenger/screens/contacts/contacts_provider.dart';
 import 'package:messenger/screens/home/home.dart';
 import 'package:messenger/services/offline/hive.db/hive_init.dart';
+import 'models/contacts_model.dart';
 import 'screens/auth/register.dart';
 import 'package:provider/provider.dart';
 import 'screens/auth/auth_provider.dart';
@@ -40,8 +42,10 @@ class MyApp extends StatelessWidget {
         ChangeNotifierProvider<AuthProvider>(
           create: (_) => AuthProvider(),
         ),
-        ChangeNotifierProvider(create: (_) => ContactProvider()),
-        FutureProvider(
+        ChangeNotifierProvider<ContactProvider>(
+            create: (_) => ContactProvider()),
+        FutureProvider<List<List<PhoneContacts>>>(
+          initialData: [],
           create: (_) => ContactProvider().registeredAndUnregisteredContacts(),
           catchError: (context, error) {
             // MessengerError _messengerError = error as MessengerError;
@@ -49,23 +53,29 @@ class MyApp extends StatelessWidget {
             // ScaffoldMessenger.of(context).showSnackBar(
             //     SnackBar(content: Text('Error fetching Contacts')));
             print(error.toString());
+            // return ;
+            throw MessengerError(error.toString());
+          },
+          builder: (context, widget) {
+            return widget ?? Container();
           },
         ),
-        ChangeNotifierProvider(create: (_) => HomeProvider()),
-        ChangeNotifierProvider(create: (_) => ChatsProvider()),
-        ChangeNotifierProvider(create: (_) => ProfileProvider()),
+        ChangeNotifierProvider<HomeProvider>(create: (_) => HomeProvider()),
+        ChangeNotifierProvider<ChatsProvider>(create: (_) => ChatsProvider()),
+        ChangeNotifierProvider<ProfileProvider>(
+            create: (_) => ProfileProvider()),
       ],
       child: MaterialApp(
         title: 'Messenger',
         theme: ThemeData(
           primarySwatch: Colors.blue,
         ),
-        home: SharedPrefs.instance.getUserData() != null
+        home: SharedPrefs.instance.getUserData().id != null
             ? HomeScreen()
             : RegistrationScreen(),
         builder: (context, child) {
           Utils.getBlockHeightAndWidth(context);
-          return child;
+          return child!;
         },
       ),
     );
