@@ -19,16 +19,16 @@ class ChatsScreen extends HookWidget {
   ChatsScreen(this.hiveChat);
   @override
   Widget build(BuildContext context) {
-    var valueListener = useState<String>();
-    final _textController = useTextEditingController();
+    var valueListener = useState<String?>("");
+    final TextEditingController? _textController = useTextEditingController();
     final _scrollController = useScrollController();
     final _chatsProvider = Provider.of<ChatsProvider>(context);
     print(hiveChat.chatId);
     return Scaffold(
       appBar: CustomAppBar(
         context: context,
-        friendContactName: hiveChat.participants[1].userName,
-        photoUrl: hiveChat.participants[1].photoUrl,
+        friendContactName: hiveChat.participants![1].userName,
+        photoUrl: hiveChat.participants![1].photoUrl,
       ),
       body: Column(
         mainAxisAlignment: MainAxisAlignment.end,
@@ -56,7 +56,7 @@ class ChatsScreen extends HookWidget {
                   itemBuilder: (context, index) {
                     // _scrollController.position
                     bool isMe = hiveMessages[index].senderID ==
-                        hiveChat.participants[0].id;
+                        hiveChat.participants![0].id;
                     _chatsProvider.updateMessageIsRead(hiveMessages[index]);
                     return Row(
                       mainAxisAlignment: isMe
@@ -93,7 +93,7 @@ class ChatsScreen extends HookWidget {
                                         BorderRadius.all(Radius.circular(10)),
                                   ),
                                   child: Text(
-                                    hiveMessages[index].msg,
+                                    hiveMessages[index].msg!,
                                     style: TextStyle(
                                         fontSize: Utils.blockWidth * 3.5),
                                   ),
@@ -104,26 +104,19 @@ class ChatsScreen extends HookWidget {
                                 width: Utils.blockWidth * 45,
                                 child: Text(
                                   DateFormat('HH:mm a').format(DateTime(
-                                              hiveMessages[index].dateTime.year,
-                                              hiveMessages[index]
-                                                  .dateTime
-                                                  .month,
-                                              hiveMessages[index].dateTime.day,
-                                              hiveMessages[index].dateTime.hour,
-                                              hiveMessages[index]
-                                                  .dateTime
-                                                  .minute,
-                                              hiveMessages[index]
-                                                  .dateTime
-                                                  .second,
-                                              hiveMessages[index]
-                                                  .dateTime
-                                                  .millisecond,
-                                              hiveMessages[index]
-                                                  .dateTime
-                                                  .microsecond)
-                                          .toLocal()) ??
-                                      "non",
+                                          hiveMessages[index].dateTime!.year,
+                                          hiveMessages[index].dateTime!.month,
+                                          hiveMessages[index].dateTime!.day,
+                                          hiveMessages[index].dateTime!.hour,
+                                          hiveMessages[index].dateTime!.minute,
+                                          hiveMessages[index].dateTime!.second,
+                                          hiveMessages[index]
+                                              .dateTime!
+                                              .millisecond,
+                                          hiveMessages[index]
+                                              .dateTime!
+                                              .microsecond)
+                                      .toLocal()),
                                   textAlign:
                                       isMe ? TextAlign.right : TextAlign.left,
                                   style: TextStyle(
@@ -166,16 +159,16 @@ class ChatsScreen extends HookWidget {
                   ),
                 ),
                 GestureDetector(
-                  onTap: valueListener.value != null
+                  onTap: valueListener.value!.length > 0
                       ? () {
-                          String msg = valueListener.value;
+                          String? msg = valueListener.value;
                           _chatsProvider.sendMessage(
                               hiveChat: hiveChat, msg: msg);
-                          _textController.clear();
-                          valueListener.value = null;
+                          _textController!.clear();
+                          valueListener.value = "";
                         }
                       : () {
-                          print(_textController.text.length);
+                          print(_textController!.text.length);
                         },
                   child: Container(
                     width: 100,
