@@ -7,11 +7,11 @@ abstract class Offline {
 
   // SharedPReferences: User Personal info and app Data
 
-  String getString(String key) {
-    return null;
+  String? getString(String key) {
+    throw UnimplementedError();
   }
 
-  bool getBool(String key) {
+  bool? getBool(String key) {
     return false;
   }
 
