@@ -5,40 +5,43 @@ import 'package:messenger/services/manager/mqtt.manager.dart';
 import 'package:messenger/services/manager/manager.dart';
 import 'package:mqtt_client/mqtt_client.dart';
 
-class MQTThandler extends ManagerHandler<MQTTManager> {
+class MQTThandler extends ManagerHandler<MQTTManager?> {
   MQTThandler() {
-    MQTTManager _newManager = MQTTManager.getInstance("broker.emqx.io", user.id,
-        user.id.substring(0, 10), user.phoneNumbers[0].substring(0, 7));
+    MQTTManager? _newManager = MQTTManager.getInstance(
+        "broker.emqx.io",
+        user.id,
+        user.id!.substring(0, 10),
+        user.phoneNumbers![0].substring(0, 7));
     setManager(_newManager);
   }
   @override
-  MQTTManager setManager(MQTTManager newManager) {
+  MQTTManager? setManager(MQTTManager? newManager) {
     return super.setManager(newManager);
   }
 
   @override
   Future<void> connectToClient() async {
-    await manager.connectMQTTClient();
+    await manager!.connectMQTTClient();
   }
 
   @override
   void disconnectClient() {
-    manager.disconnectMQTTClient();
+    manager!.disconnectMQTTClient();
   }
 
   @override
   Future<MqttClient> login() async {
-    return await manager.login();
+    return await manager!.login();
   }
 
   @override
   Future<void> publish(String topic, Message message) async {
-    manager.publish(topic, message.toMap());
+    manager!.publish(topic, message.toMap());
   }
 
   @override
-  Future<bool> subscribe(String topic) async {
-    return manager.subscribe(topic);
+  Future<bool> subscribe(String? topic) async {
+    return manager!.subscribe(topic);
   }
 
   @override
@@ -46,5 +49,5 @@ class MQTThandler extends ManagerHandler<MQTTManager> {
     super.dispose();
   }
 
-  Stream<Map<String, dynamic>> get messageController => manager.messageStream;
+  Stream<Map<String, dynamic>?> get messageController => manager!.messageStream;
 }
