@@ -14,8 +14,8 @@ extension dateFormat on String {
 }
 
 extension capitalizeString on String {
-  String capitalize() {
-    var listOfChar = this.split("");
+  String? capitalize() {
+    List<String> listOfChar = this.split("");
     listOfChar.replaceRange(0, 1, [listOfChar.first.toUpperCase()]);
     return listOfChar.join("");
   }
