@@ -4,4 +4,4 @@ import 'package:flutter/material.dart';
 typedef VoidUserCallBack = void Function(User);
 typedef VoidStringCallBack = void Function(String);
 typedef DoubleValueListenableWidgetBuilder<T, E> = Widget Function(
-    BuildContext context, T value1, E value2, Widget child);
+    BuildContext context, T? value1, E? value2, Widget? child);
